backend/requirements.txt
torch
opencv-python
numpy
Pillow
fastapi
uvicorn
fpdf
python-multipart
gtts
